reorg mux and slice
