+++ magma/primitives/__init__.py
@@ -1,3 +1,4 @@
 from magma.primitives.lut import LUT
-from magma.primitives.mux import Mux
+from magma.primitives.mux import Mux, mux
 from magma.primitives.register import Register
+from magma.primitives.slice import slice
